requirements.txt
celery[redis]==5.2.7
dj-database-url==1.0.0
Django==4.1.3
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-google-tag-manager==0.0.5
django-hut-theme==0.11
django-robots==5.0
django-settings-export==1.2.1
flower==1.2.0
gunicorn==20.1.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
whitenoise==6.2.0


# Private requirements
--extra-index-url https://${FURY_AUTH}:@pypi.fury.io/hutfortytwo/
django-hut-theme
